move entities out of post.go into seperate files
