+++ author.go
@@ -1,0 +1,7 @@
+package analogdb
+
+import "context"
+
+type AuthorService interface {
+	FindAuthors(ctx context.Context) ([]string, error)
+}
