+++ scrape.go
@@ -1,0 +1,7 @@
+package analogdb
+
+import "context"
+
+type ScrapeService interface {
+	KeywordUpdatedPostIDs(ctx context.Context) ([]int, error)
+}
